Fix issues with object store usage
