+++ src/bulkApi/config/swagger.yaml
@@ -326,17 +326,7 @@ definitions:
         description: Id of transfer
         x-format:
           guid: true
-      payeeFsp:
-        type: string
-        description: Financial Service Provider of Payee
-        minLength: 1
-        maxLength: 32
-      payerFsp:
-        type: string
-        description: Financial Service Provider of Payer
-        minLength: 1
-        maxLength: 32
-      amount:
+      transferAmount:
         $ref: '#/definitions/Amount'
       ilpPacket:
         type: string
@@ -351,21 +341,13 @@ definitions:
         pattern: '^[A-Za-z0-9-_]{43}$'
         x-convert:
           trim: true
-      expiration:
-        type: string
-        description: When the transfer expires
-        pattern: >-
-          ^(?:[1-9]\d{3}-(?:(?:0[1-9]|1[0-2])-(?:0[1-9]|1\d|2[0-8])|(?:0[13-9]|1[0-2])-(?:29|30)|(?:0[13578]|1[02])-31)|(?:[1-9]\d(?:0[48]|[2468][048]|[13579][26])|(?:[2468][048]|[13579][26])00)-02-29)T(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:(\.\d{3}))(?:Z|[+-][01]\d:[0-5]\d)$
       extensionList:
         $ref: '#/definitions/extensionList'
     required:
       - transferId
-      - payeeFsp
-      - payerFsp
-      - amount
+      - transferAmount
       - ilpPacket
       - condition
-      - expiration
   TransferFulfilment:
     type: object
     properties:
