+++ src/bulkApi/handlers/bulkTransfers.js
@@ -4,6 +4,7 @@ const TransferService = require('../../domain/transfer')
 const Logger = require('@mojaloop/central-services-shared').Logger
 const Boom = require('boom')
 const { BulkTransferModel } = require('../models/bulkTransfers/bulkModels')
+const Util = require('../../lib/util')
 
 /**
  * Operations on /bulkTransfers
@@ -19,10 +20,11 @@ module.exports = {
   post: async function postBulkTransfers (request, h) {
     try {
       Logger.debug('create::payload(%s)', JSON.stringify(request.payload))
+      let { bulkTransferId, bulkQuoteId, payerFsp, payeeFsp, expiration, extensionList } = request.payload
+      let hash = Util.createHash(JSON.stringify(request.payload))
       let newBulk = new BulkTransferModel(Object.assign({}, { headers: request.headers }, request.payload))
-      let savedBulk = await newBulk.save()
-      let { bulkTransferId, bulkQuoteId, payerFsp, payeeFsp, expiration, extensionList, status } = savedBulk
-      let message = { bulkTransferId, bulkQuoteId, payerFsp, payeeFsp, expiration, extensionList, status }
+      let { status } = await newBulk.save()
+      let message = { bulkTransferId, bulkQuoteId, payerFsp, payeeFsp, expiration, extensionList, hash, status }
       await TransferService.bulkPrepare(request.headers, message, request.dataUri)
       return h.response().code(202)
     } catch (err) {
