+++ src/bulkApi/server.js
@@ -4,6 +4,8 @@ const Hapi = require('hapi')
 const HapiOpenAPI = require('hapi-openapi')
 const Path = require('path')
 const Mongoose = require('./lib/mongodb/db').Mongoose
+const Boom = require('@hapi/boom')
+const Logger = require('@mojaloop/central-services-shared').Logger
 
 const connectMongoose = async () => {
   let db = await Mongoose.connect(`mongodb://localhost:27017/bulk_transfers`, {
@@ -14,7 +16,16 @@ const connectMongoose = async () => {
 
 const init = async function (options) {
   const { port } = options
-  const server = new Hapi.Server({ port })
+  const server = new Hapi.Server({
+    port,
+    routes: {
+      validate: {
+        failAction: async (request, h, err) => {
+          throw Boom.boomify(err)
+        }
+      }
+    }
+  })
   let db = await connectMongoose()
   server.app.db = db
   await server.register({
@@ -24,16 +35,15 @@ const init = async function (options) {
       handlers: Path.resolve(__dirname, './handlers')
     }
   })
-
   await server.start()
-
   return server
 }
 
 const initialize = (options) => {
   init(options).then((server) => {
     server.plugins.openapi.setHost(server.info.host + ':' + server.info.port)
-    server.log(['info'], `Server running on ${server.info.host}:${server.info.port}`)
+    Logger.info(`Server running at: ${server.info.host}:${server.info.port}`)
+    server.log(['info'], `Server running at: ${server.info.host}:${server.info.port}`)
     return server
   })
 }
